% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_find_tests}
\alias{bh_find_tests}
\title{Find RawData files}
\usage{
bh_find_tests(study_root, raw_regex = "")
}
\arguments{
\item{study_root}{the path where to start searching for specimen RawData files.}

\item{raw_regex}{defines what files to read. Defaults to all RawData files.}
}
\value{
A list with the full path name to each specimen file under study_root.
}
\description{
Starting from \code{study_root}, grab the names of everything that looks like an Instron
specimen as defined by \code{raw_regex}. The default is all RawData files under \code{study_root}.
If only a sub-set of the sample files are needed, supply a regex to select just those.
for example, if there are several samples under study.root and we only want those for "bone",
use something like
  bh_find_tests(study_root, raw_regex = ".*bone.*RawData.*\\.csv")
Or better, gather all the file names and filter the returned list.
}
