% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_get_headers}
\alias{bh_get_headers}
\title{Get Bluehill Headers}
\usage{
bh_get_headers(samples)
}
\arguments{
\item{samples}{A \code{data.table} with the pathname of each specimen file in column \code{filename}.}
}
\value{
A \code{data.table} with the headers (if any) from each specimen file
  read using \code{bh_read_header}.
}
\description{
Given a \code{data.table} containing the path names of the desired RawData files
in column \code{filename}, return all the headers for each specimen in a \code{data.table}
keyed by filename and the header names (in column \code{var}).
}
