% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.2.R, R/bluehill_functions.R
\name{bh_get_labels}
\alias{bh_get_labels}
\title{Get important sample labels from headers}
\usage{
bh_get_labels(headers)

bh_get_labels(headers)
}
\arguments{
\item{headers}{A \code{data.table} possibly containing a \code{blank_row} and \code{Specimen Label}.}

\item{headers}{A \code{data.table} possibly containing a \code{blank_row} and \code{Specimen Label}.}
}
\value{
a named list containing the \code{blank_row}, \code{Specimen Label} and \code{filename} for each specimen.

a named list containing the \code{blank_row}, \code{Specimen Label} and \code{filename} for each specimen.
}
\description{
Extracts the \code{blank_row} and \code{Specimen Label} parts a specimen header.

Extracts the \code{blank_row} and \code{Specimen Label} parts a specimen header.
}
