% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_invert}
\alias{bh_invert}
\title{Change sign of compressive test}
\usage{
bh_invert(study_data)
}
\arguments{
\item{study_data}{A \code{data.table} containing at least columns for \code{Load} and \code{Extension}.}
}
\value{
The same \code{data.table} with \code{ -1.0 * Load} and \code{-1.0 * Extension}
}
\description{
Compression is defined as negative, so compressive tests have negative Load and Extension.
Bluehill provides \code{Compressive Load} and \code{Compressive Extension}, but these are
not in every test, so this function inverts \code{Load} and \code{Extension} so that every
test will be the same way up (i.e. move from low to high extension and low to high load).
}
