% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_label_cycles}
\alias{bh_label_cycles}
\title{Label cycles & segments using robust_peaks}
\usage{
bh_label_cycles(study, channel = "Extension", span = 5)
}
\arguments{
\item{study}{a BlueHill study to mark with peaks and cycles.}

\item{channel}{the data channel to search for peaks.}

\item{span}{size of window to use when looking for peaks. Default is 5.
Larger spans are less sensitive to noise, but effectively smooth the series.}
}
\value{
a list of cycle numbers cycle = 1:n.cycles
  and segments seg = rep(c("load","unload"))
  both padded to the length of the original series
}
\description{
Given a data series, break into cycles (at each trough)
break each cycle into load (trough -> peak) and
unload (peak -> trough) segments, using turning points
from robust_peaks.
As tests start at a trough and go to a peak, the testing
direction is found by checking sign of the first peak/trough.
Only works on "positive going" data, that is where the Load and
Extension increase from start to peak.
In other words, compressive studies must be inverted first.
}
