% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.2.R
\name{bh_label_cycles_old}
\alias{bh_label_cycles_old}
\title{Label cycles & segments}
\usage{
bh_label_cycles_old(specimen, channel = "Extension", span = 3)
}
\arguments{
\item{specimen}{a \code{data.table} from BlueHill with the channels to search.}

\item{channel}{which channel to look in for peaks.
Defaults to "Extension", which is usually the cleanest channel.}

\item{span}{size of window to use when looking for peaks. Defaults to 3.
Larger spans are less sensitive to noise, but effectively smooth the series.}
}
\value{
the input \code{data.table} augmented with columns for
  \code{cycle} the cycle number = 1:n.cycles and
  \code{segment} the segnment = rep(c("load","unload"))
  \code{peaks} = (-1, 0, +1) marking the location and direction of any peaks.
}
\description{
Given a data series, break into cycles (at each trough)
break each cycle into load (trough -> peak) and
unload (peak -> trough) segments, using turning points
from \code{peaksign}.
As tests start at a trough and go to a peak, the testing
direction is found by checking sign of the first peak/trough.
}
