% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_make_compressive}
\alias{bh_make_compressive}
\title{Change sign of compressive test}
\usage{
bh_make_compressive(sample_data)
}
\arguments{
\item{sample_data}{A \code{data.table} containing at least columns for \code{Load} and \code{Extension}.}
}
\value{
The same \code{data.table} with \code{ -1.0 * Load} and \code{-1.0 * Extension}
}
\description{
Compression is defined as negative, so compressive tests have negative Load and Extension.
This can look odd when plotted, so Bluehill provides \code{Compressive Load} and \code{Compressive Extension}
This function inverts \code{Load} and \code{Extension}
}
