% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\docType{data}
\name{bh_min_cols}
\alias{bh_min_cols}
\title{Minimal Bluehill columns}
\format{An object of class \code{character} of length 3.}
\usage{
bh_min_cols
}
\description{
A constant naming the minimal column names for a Bluehill test:
Time, Extension, Load
}
\keyword{datasets}
