% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_min_columns}
\alias{bh_min_columns}
\title{Minimal Bluehill columns}
\usage{
bh_min_columns()
}
\value{
a character vector = c("Time", "Extension", "Load")
}
\description{
Returns the minimal column names for a Bluehill test:
Time, Extension, Load
}
