% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_modulus}
\alias{bh_modulus}
\title{Calculate the slope(modulus) of a data series using the Bluehill algorithm
Essentially:
  divide the span from \code{min(y)} to \code{max(y)} into 5 segments
  calculate the slope of each segment
  find the pair of segments with the highest average slope
  combine those segments and recalcuate the slope for the both together}
\usage{
bh_modulus(x, y, intervals = 5)
}
\arguments{
\item{x}{a vector of ordinate data (often time or extension)}

\item{y}{a vector of abcissa data}
}
\value{
the highest average slope(modulus) of \code{y/x}
}
\description{
Calculate the slope(modulus) of a data series using the Bluehill algorithm
Essentially:
  divide the span from \code{min(y)} to \code{max(y)} into 5 segments
  calculate the slope of each segment
  find the pair of segments with the highest average slope
  combine those segments and recalcuate the slope for the both together
}
\examples{
stiffness <- bh_modulus(Extension, Load)
}
