% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_normalise}
\alias{bh_normalise}
\title{Make all tests positive going}
\usage{
bh_normalise(study_data)
}
\arguments{
\item{study_data}{A \code{data.table} containing at least columns \code{Load} and \code{Extension}.}
}
\value{
The same \code{data.table} with \code{Load} and \code{Extension} always going from low to high
}
\description{
Compression is defined as negative, so compressive tests have negative Load and Extension.
Bluehill provides \code{Compressive Load} and \code{Compressive Extension}, but these are
not in every test, so this function checks if \code{Load} is going negative
(i.e. a compressive test) and if so inverts \code{Load} and \code{Extension} so
that the test will move from low to high extension and low to high load.
}
