% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.2.R
\docType{data}
\name{bh_options}
\alias{bh_options}
\title{Minimal Bluehill columns}
\format{An object of class \code{list} of length 6.}
\usage{
bh_options
}
\description{
A constant naming the minimal column names for a Bluehill test:
Time, Extension, Load
}
\keyword{datasets}
