% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_read_data}
\alias{bh_read_data}
\title{Read the data from a RawData file
RawData files can contain header rows that describe the test, followed by a blank line
and then the measured channels in columns. This function reads the data channels, starting
after \code{blank_row} to skip any headers. Data channels have two line headers, with the
channel names in the first line and units in the second. This function discards channel units.
If \code{min_results} = TRUE then only the Time, Extension and Load channels are returned.
By default all the channels in \code{filename} are read.}
\usage{
bh_read_data(filename, blank_row = 0, select_cols = NULL)
}
\arguments{
\item{filename}{the RawData file to process.}

\item{blank_row}{line number of the end of any header. Defaults to 0 (no header).}

\item{select_cols}{a list of columns to return by number (or all if NULL). Defaults to NULL.}
}
\value{
A \code{data.table} with the measured channels in columns and the channel names as column names.
}
\description{
Read the data from a RawData file
RawData files can contain header rows that describe the test, followed by a blank line
and then the measured channels in columns. This function reads the data channels, starting
after \code{blank_row} to skip any headers. Data channels have two line headers, with the
channel names in the first line and units in the second. This function discards channel units.
If \code{min_results} = TRUE then only the Time, Extension and Load channels are returned.
By default all the channels in \code{filename} are read.
}
