% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_read_header}
\alias{bh_read_header}
\title{Read any headers in a RawData file}
\usage{
bh_read_header(filename)
}
\arguments{
\item{filename}{the RawData file to process}
}
\value{
A \code{data.table} with any parameter : value pairs and
  the special header \code{blank_row} indicating the end of the header,
  or 0 if there was no header.
}
\description{
RawData files can contain header rows that describe the test as
parameter : value pairs. This function reads any parameters into
a \code{data.table} with columns for "type", "var", "value" and "units".
A special var \code{blank_row} of type \code{information} is added at
the end with the row number of the end of the header, or 0 if there was
no header in the file. This can be passed to \code{bh_read_data}.
}
