% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_read_raw}
\alias{bh_read_raw}
\title{Read a BlueHill RawData file}
\usage{
bh_read_raw(filename, min_results = TRUE, use_label = TRUE,
  use_filename = FALSE, headers = FALSE)
}
\arguments{
\item{filename}{the name of the RawData file to process.}

\item{min_results}{if \code{TRUE} only return three columns:
(Time, Load, Position). Defaults to \code{TRUE}.}

\item{use_label}{if \code{TRUE} add any \code{label} field in the header
as a column in the output. Defaults to \code{TRUE}.}

\item{use_filename}{if \code{TRUE} use the filename as the specimen_ID
otherwise add a unique numeric specimen_ID for each file.
Defaults to \code{FALSE}.}

\item{headers}{if \code{TRUE} return a \code{data.table} containing any
parameter:value pairs and \code{blank_row} indicating the end of the header.
\code{blank_row} = 0 if there were no headers. Defaults to \code{FALSE}.}
}
\value{
if \code{headers == TRUE} a list with data = a \code{data.table} holding
  the raw data and header = a \code{data.table} holding headers
   if \code{headers == FALSE} just the \code{data.table} holding the raw data
}
\description{
RawData files can contain optional header rows that describe the test as
parameter: value pairs, followed by a blank row, then the recorded data.
}
\details{
The data part has two rows of column header: variable names in the first row,
then units in the second. Subsequent rows have the data in columns.

This function reads any parameters into a \code{data.table} with columns for
"type", "var", "value" and "units" A special header is added to describe the
location of the end of the header as \code{information, blank_row, n}
where\code{n} is the row number of the blank line that marks the end of the
header (0 is there was no header). This can be passed to \code{bh_read_data}
to indicate where the data starts.
}
