% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.2.R
\name{bh_read_study}
\alias{bh_read_study}
\title{read a set of tests (specimens), given their filenames}
\usage{
bh_read_study(filenames)
}
\value{
a list containing two \code{data.table}s
  \code{headers} holding any headers, indexed by testID and
  \code{data}  containing all the data for all of the files specified.
  Each specimen has a unique ID in column testID, the same as \code{headers}.
}
\description{
read a set of tests (specimens), given their filenames
}
