% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.2.R
\name{bh_set_options}
\alias{bh_set_options}
\title{Set __bluer__ options}
\usage{
bh_set_options(raw_regex = ".*RawData.*\\\\.csv", max_header = 100,
  min_columns = c("Time", "Extension", "Load"),
  id_regex = file.path(".*",
  "(.+).is_.+_RawData.*Specimen_RawData[\\\\_\\\\.](\\\\d+)\\\\.csv"),
  id_parts = c("filename", "sample", "specimen"))
}
\description{
Provides an interface to set global options of processing BlueHill files.
}
