% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueAlert.R
\name{blueAlert}
\alias{blueAlert}
\title{Create a Boostrap 4 blue alert}
\usage{
blueAlert(..., icon = NULL, status = "default", closable = TRUE)
}
\arguments{
\item{...}{Alert content.}

\item{icon}{Alert icon. Expect \link{blueIcon} or \link[shiny]{icon}.}

\item{status}{Alert status. See \url{https://demos.creative-tim.com/blue-design-system/docs/components/alerts.html}.}

\item{closable}{Whether to make the alert closable or not. TRUE by default.}
}
\description{
Build an blue alert
}
\examples{
if (interactive()) {
 library(blueR)
 blueAlert(
  icon = blueIcon("basket"),
  status = "danger",
  "This is an alert",
  closable = TRUE
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
