% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueBadge.R
\name{blueBadge}
\alias{blueBadge}
\title{Create a Boostrap 4 badge}
\usage{
blueBadge(text = NULL, src = NULL, pill = FALSE, status = "default")
}
\arguments{
\item{text}{Badge text.}

\item{src}{Badge external link.}

\item{pill}{Pill mode. FALSE by default.}

\item{status}{Badge status. See \url{https://demos.creative-tim.com/blue-design-system/docs/components/badge.html}.}
}
\description{
Build an blue badge
}
\examples{
if (interactive()) {
 library(blueR)
 blueBadge(
   text = "My badge",
   src = "https://www.google.com",
   pill = FALSE,
   status = "success"
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
