% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{blueBlur}
\alias{blueBlur}
\title{Create a Boostrap 4 blur effect}
\usage{
blueBlur(..., text = NULL, text_color = NULL)
}
\arguments{
\item{...}{Tag to be treated by the blur effect.}

\item{text}{Text that appears on hover.}

\item{text_color}{Text color. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}
}
\description{
Build an blue blur effect
}
\note{
Not compatible with blueDash
}
\examples{
if (interactive()) {
 library(blueR)
 blueBlur(
  blueImage(
   src = "inst/images/imac.svg",
   floating = TRUE
  )
 ) 
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
