% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueCarousel.R
\name{blueCarouselItem}
\alias{blueCarouselItem}
\title{Create a carousel item}
\usage{
blueCarouselItem(src = NULL, active = FALSE, mode = "img")
}
\arguments{
\item{src}{Image url or path.}

\item{active}{Whether the item is active or not.}

\item{mode}{Item mode: "img" by default but also "iframe" or "video".}
}
\description{
Build an blue carousel item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
