% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{blueCascade}
\alias{blueCascade}
\title{Create a Boostrap 4 cascade effect}
\usage{
blueCascade(...)
}
\arguments{
\item{...}{Slot for \link{blueCascadeItem}. Exactly 13 items.}
}
\description{
Build an blue cascade effect
}
\note{
You must pass cascade = TRUE to blueSection
}
\examples{
if (interactive()) {
 library(blueR)
 blueSection(
  size = "lg",
  status = "default",
  cascade = TRUE,
  
  blueH1(display = 3, "BlueR Cascade Effect") \%>\% 
    bluePadding(orientation = "l", value = 5) \%>\%
    bluePadding(orientation = "b", value = 5) \%>\%
    blueTextColor(color = "white"),
  
  blueCascade(
    blueCascadeItem(name = "diamond", src = "https://www.google.com"),
    blueCascadeItem(name = "album-2", size = "sm"),
    blueCascadeItem(name = "app", size = "sm"),
    blueCascadeItem(name = "atom", size = "sm"),
    blueCascadeItem(name = "bag-17", src = "https://www.google.com"),
    blueCascadeItem(name = "bell-55"),
    blueCascadeItem(name = "credit-card"),
    blueCascadeItem(name = "briefcase-24", size = "sm", "https://www.google.com"),
    blueCascadeItem(name = "building", size = "sm"),
    blueCascadeItem(name = "button-play", size = "sm"),
    blueCascadeItem(name = "calendar-grid-58"),
    blueCascadeItem(name = "camera-compact"),
    blueCascadeItem(name = "chart-bar-32")
  )
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
