% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{blueCascadeItem}
\alias{blueCascadeItem}
\title{Create a Boostrap 4 cascade item}
\usage{
blueCascadeItem(name, size = NULL, src = NULL)
}
\arguments{
\item{name}{Icon name.}

\item{size}{Icon size: NULL, "sm" or "lg".}

\item{src}{Optional external url, link, ...}
}
\description{
Build an blue cascade item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
