% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueColumn.R
\name{blueColumn}
\alias{blueColumn}
\title{Create a Boostrap 4 column}
\usage{
blueColumn(..., width = NULL, center = FALSE, offset = NULL)
}
\arguments{
\item{...}{Any UI element.}

\item{width}{Optional. Column width between 1 and 12.}

\item{center}{Whether to center column elements or not. FALSE by default.}

\item{offset}{Column offset. NULL by default. Importantly, offset + width cannot exceed 12!}
}
\description{
Build an blue column
}
\examples{
if (interactive()) {
 library(blueR)
 blueColumn(
  width = 4,
  blueAlert(
    icon = "basket",
    status = "danger",
    "This is an alert",
    closable = TRUE
  )
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
