% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueNavbar.R
\name{blueDropdown}
\alias{blueDropdown}
\title{Create a Boostrap 4 blue dropdown menu}
\usage{
blueDropdown(..., name, size = NULL)
}
\arguments{
\item{...}{Slot for \link{blueDropdownItem}.}

\item{name}{Dropdown menu name.}

\item{size}{Dropdown size: NULL or "xl".}
}
\description{
Build an blue dropdown menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
