% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueNavbar.R
\name{blueDropdownItem}
\alias{blueDropdownItem}
\title{Create a Boostrap 4 blue dropdown item}
\usage{
blueDropdownItem(
  name = NULL,
  description = NULL,
  src = NULL,
  icon = NULL,
  status = NULL
)
}
\arguments{
\item{name}{Item name.}

\item{description}{Item description if any.}

\item{src}{HTML target page.}

\item{icon}{Item icon of any. Expect \link{blueIcon} or \link[shiny]{icon}.}

\item{status}{Icon and name color status. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}
}
\description{
Build an blue dropdown item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
