% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueFooter.R
\name{blueFooter}
\alias{blueFooter}
\title{Create a Boostrap 4 blue footer}
\usage{
blueFooter(..., has_card = FALSE, status = NULL, gradient = FALSE)
}
\arguments{
\item{...}{Footer content.}

\item{has_card}{Enable card rendering in the footer. FALSE by default.}

\item{status}{Footer status color. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}

\item{gradient}{Whether to apply a gradient. FALSE by default.}
}
\description{
Build an blue footer
}
\examples{
if(interactive()){
 library(blueR)
 blueFooter(
  has_card = FALSE,
  blueContainer(
    size = "lg",
    blueRow(
      blueColumn(
        width = 6,
        blueIconWrapper(
          iconTag = blueIcon("atom"),
          size = "lg",
          status = "success",
          shadow = TRUE,
          hover_shadow = TRUE
        ),
        blueH1(
          display = 3,
          "Insert anything in the footer"
        ),
        blueLead(
          "The Arctic Ocean freezes every winter and much of 
          the sea-ice then thaws every summer, and that process 
          will continue whatever"
        )
        ),
      blueColumn(
        width = 6,
        blueCarousel(
          width = 12,
          id = "carousel2",
          blueCarouselItem(
            src = "http://www.72pxdesigns.com/wp-content/uploads/2017/06/preview-814162.png",
            active = TRUE
          ),
          blueCarouselItem(
            src = "http://www.72pxdesigns.com/wp-content/uploads/2017/06/preview-814162.png",
            active = FALSE
          )
        )
      )
    )
   )
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
