% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{blueH1}
\alias{blueH1}
\title{Create a custim h1 tag}
\usage{
blueH1(..., display)
}
\arguments{
\item{...}{Any text.}

\item{display}{value between 1 and 4.}
}
\description{
A custom h1 tag
}
\examples{
if (interactive()) {
 library(blueR)
 blueH1(display = 3, "BlueR elements")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
