% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueIcon.R
\name{blueIcon}
\alias{blueIcon}
\title{Create a Nucleo icon}
\usage{
blueIcon(name, color = NULL)
}
\arguments{
\item{name}{icon name.}

\item{color}{icon color. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}
}
\description{
Build an blue Nucleo icon
}
\examples{
if (interactive()) {
 library(blueR)
 blueIcon("atom")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
