% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueIcon.R
\name{blueIconWrapper}
\alias{blueIconWrapper}
\title{Create a Nucleo icon wrapper}
\usage{
blueIconWrapper(
  iconTag,
  circle = TRUE,
  size = NULL,
  status = "default",
  gradient_color = NULL,
  shadow = TRUE,
  hover_shadow = FALSE
)
}
\arguments{
\item{iconTag}{Slot for \link{blueIcon}}

\item{circle}{Wrapper's shape. TRUE by default.}

\item{size}{Wrapper size. "sm", "md" or "lg".}

\item{status}{Wrapper color. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}

\item{gradient_color}{icon gradient background color.}

\item{shadow}{Whether to apply a shadow effet. TRUE by default.}

\item{hover_shadow}{Only if shadow is TRUE. Whether to enlarge the shadow on hover. FALSE by default.}
}
\description{
Build an blue Nucleo icon wrapper
}
\examples{
if (interactive()) {
 library(blueR)
 blueIconWrapper(
  iconTag = blueIcon("atom"),
  size = "lg",
  status = "danger",
  shadow = TRUE,
  hover_shadow = TRUE
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
