% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluePage.R
\name{blueInstall}
\alias{blueInstall}
\title{Install blue assets}
\usage{
blueInstall(path = getwd())
}
\arguments{
\item{path}{Where to store assets. By default, getwd().}
}
\description{
Install blue assets
}
\note{
blue assets are stored in the inst folder similarly as the package folder.
}
\examples{
if(interactive()){
 library(blueR)
 blueInstall()
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
