% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{blueLead}
\alias{blueLead}
\title{Create a custom p tag}
\usage{
blueLead(...)
}
\arguments{
\item{...}{Any text.}
}
\description{
A custom p tag
}
\examples{
if (interactive()) {
 library(blueR)
 blueLead("Badges")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
