% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{blueMuted}
\alias{blueMuted}
\title{Create a muted text}
\usage{
blueMuted(...)
}
\arguments{
\item{...}{Any text.}
}
\description{
Apply the class muted to a text
}
\examples{
if (interactive()) {
 library(blueR)
 blueMuted(blueLead("Badges"))
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
