% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueNavbar.R
\name{blueNavItem}
\alias{blueNavItem}
\title{Create a Boostrap 4 blue navbar item}
\usage{
blueNavItem(name = NULL, src = NULL, icon = NULL, tooltip = NULL)
}
\arguments{
\item{name}{Item name.}

\item{src}{HTML target page.}

\item{icon}{Item icon of any. Expect \link{blueIcon} or \link[shiny]{icon}.}

\item{tooltip}{Text to display when the item is hovered.}
}
\description{
Build an blue navbar item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
