% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueNavbar.R
\name{blueNavMenu}
\alias{blueNavMenu}
\title{Create a Boostrap 4 blue navbar menu}
\usage{
blueNavMenu(..., side = "left")
}
\arguments{
\item{...}{Slot for \link{blueNavItem} and \link{blueDropdown}.}

\item{side}{Nav menu side. "left by default" but can also be "right".}
}
\description{
Build an blue navbar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
