% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{bluePadding}
\alias{bluePadding}
\title{Change tag padding}
\usage{
bluePadding(tag, orientation, value)
}
\arguments{
\item{tag}{Tag to be treated.}

\item{orientation}{Padding side. See \url{https://getbootstrap.com/docs/4.1/utilities/spacing/}.}

\item{value}{Padding value. See \url{https://getbootstrap.com/docs/4.1/utilities/spacing/}.}
}
\description{
Change tag padding
}
\examples{
if (interactive()) {
 library(blueR)
 bluePadding(tag = h1("test"), orientation = NULL, value = 5)
 bluePadding(tag = h1("test"), orientation = "x", value = 2)
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
