% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluePage.R
\name{bluePage}
\alias{bluePage}
\title{Create a Boostrap 4 blue page}
\usage{
bluePage(
  ...,
  title = NULL,
  description = NULL,
  author = NULL,
  navbar = NULL,
  footer = NULL,
  favicon = NULL,
  analytics = NULL
)
}
\arguments{
\item{...}{Body content}

\item{title}{App title.}

\item{description}{Purpose.}

\item{author}{Author.}

\item{navbar}{Navbar.}

\item{footer}{Footer.}

\item{favicon}{Website favicon. The png must be located in inst/images.}

\item{analytics}{Website analytics such as Hotjar or google analytics. 
Must be wrapped in tagList or list. Moreover the script must be contained
in a <script></script> tag: if it is not already the case, use tags$script.}
}
\description{
Build an blue page
}
\examples{
if(interactive()){
 library(blueR)
 bluePage(
  title = "BlueR Static Template",
  author =  "Somebody",
  description = "HTML Static Template",
  navbar = blueNavbar(id = "navbar"),
  footer = blueFooter(),
  # main content
  blueSection(),
  blueSection(),
  blueSection(),
  blueSection(),
  blueSection()
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
