% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluePage.R
\name{bluePageTemplate}
\alias{bluePageTemplate}
\title{Create an HTML version of the bluePage}
\usage{
bluePageTemplate(filename = "index", path = getwd(), bluePage, view = TRUE)
}
\arguments{
\item{filename}{HTML filename for instance, index.html.}

\item{path}{Where to store the saved file. By default, getwd().}

\item{bluePage}{Slot for \link{bluePage}.}

\item{view}{Whether to preview the page in a web browser. TRUE by default.}
}
\description{
Create an HTML version of the bluePage
}
\note{
Do not forget to copy the inst folder of the package to the 
root of your website folder.
}
\examples{
if(interactive()){
 library(blueR)
 
 # generate the page
 example <- bluePage(
  title = "BlueR Static Template",
  author =  "Somebody",
  description = "HTML Static Template",
  navbar = blueNavbar(id = "navbar"),
  footer = blueFooter(),
  # main content
  blueSection(),
  blueSection(),
  blueSection(),
  blueSection(),
  blueSection()
 )
 
 # create the path
 path <- getwd()
 
 # generate the static page
 bluePageTemplate(filename = "example", path = path, bluePage = example)
 
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
