% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluePagination.R
\name{bluePagination}
\alias{bluePagination}
\title{Build an blue pagination container}
\usage{
bluePagination(..., size = NULL, align = NULL)
}
\arguments{
\item{...}{Slot for \link{bluePaginationItem}.}

\item{size}{Pagination size: NULL, "sm" or "lg".}

\item{align}{Pagination alignement. NULL (left), "center" or "end" (right).}
}
\description{
Build an blue pagination container
}
\note{
align will not work if you embed the bluePagination in an blueRow.
}
\examples{
if(interactive()){
 library(blueR)
 bluePagination(
  size = "lg",
  align = "center",
  bluePaginationItem(
    name = 1,
    src = "test.html"
  ),
  bluePaginationItem(
    name = 2,
    src = "https://www.google.com"
  )
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
