% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluePagination.R
\name{bluePaginationItem}
\alias{bluePaginationItem}
\title{Build an blue pagination item}
\usage{
bluePaginationItem(name, src = NULL)
}
\arguments{
\item{name}{Item name. Better to use numerics such as 1, 2, 3, ...}

\item{src}{Item target. Another html page for instance.}
}
\description{
Build an blue pagination item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
