% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-effects.R
\name{bluePersp}
\alias{bluePersp}
\title{Create a Boostrap 4 perspective effect}
\usage{
bluePersp(..., side)
}
\arguments{
\item{...}{Tag to be treated by the persp effect.}

\item{side}{Perspective side effect. "right" or "left".}
}
\description{
Build an blue perspective effect
}
\note{
Disabled on small screens (not a bug)
}
\examples{
if (interactive()) {
 library(blueR)
 bluePersp(blueCard(), side = "right")
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
