% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueProfile.R
\name{blueProfile}
\alias{blueProfile}
\title{Create a Boostrap 4 profile card}
\usage{
blueProfile(
  ...,
  title = NULL,
  subtitle = NULL,
  src = NULL,
  url = NULL,
  url_1 = NULL,
  url_2 = NULL,
  stats
)
}
\arguments{
\item{...}{Any element.}

\item{title}{Profile title.}

\item{subtitle}{Profile subtitle.}

\item{src}{Profile image path or url.}

\item{url}{Main url.}

\item{url_1}{First social link.}

\item{url_2}{Second social link.}

\item{stats}{Stats items. See \link{blueProfileStats}.}
}
\description{
Build an blue profile card
}
\note{
Do not work on blueR. Only for blue Dash
}
\examples{
if(interactive()){
 library(blueR)
 blueProfile(
  title = "John",
  subtitle = "Japan, Kagoshima",
  src = "https://image.flaticon.com/icons/svg/1006/1006540.svg",
  url = "https://www.google.com",
  url_1 = "https://www.google.com",
  url_2 = "https://www.google.com",
  stats = blueProfileStats(
    blueProfileStat(
      value = 22,
      description = "Friends"
    ),
    blueProfileStat(
      value = 10,
      description = "Photos"
    ),
    blueProfileStat(
      value = 89,
      description = "Comments"
    )
  ),
  "An artist of considerable range, Ryan — 
  the name taken by Melbourne-raised, 
  Brooklyn-based Nick Murphy — writes, 
  performs and records all of his own music, 
  giving it a warm, intimate feel with a solid 
  groove structure. An artist of considerable 
  range."
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
