% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueProfile.R
\name{blueProfileStat}
\alias{blueProfileStat}
\title{Create a Boostrap 4 profile card stat item}
\usage{
blueProfileStat(value, description)
}
\arguments{
\item{value}{Stat value.}

\item{description}{Stat description.}
}
\description{
Build an blue profile card stat item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
