% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueProfile.R
\name{blueProfileStats}
\alias{blueProfileStats}
\title{Create a Boostrap 4 profile card stats container}
\usage{
blueProfileStats(...)
}
\arguments{
\item{...}{Slot for \link{blueProfileStat}.}
}
\description{
Build an blue profile card stats container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
