% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueProgress.R
\name{blueProgress}
\alias{blueProgress}
\title{Create a progress bar}
\usage{
blueProgress(value, text = NULL, status = "primary")
}
\arguments{
\item{value}{Progress value between 0 and 100.}

\item{text}{Label text.}

\item{status}{Progress status color. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}
}
\description{
Build an blue progress bar
}
\examples{
if (interactive()) {
 library(blueR)
 blueProgress(value = 10, status = "danger", text = "Custom Text")
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
