% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueRow.R
\name{blueRow}
\alias{blueRow}
\title{Create a Boostrap 4 row}
\usage{
blueRow(..., center = FALSE)
}
\arguments{
\item{...}{Any UI element.}

\item{center}{Whether to center row elements or not. FALSE by default.}
}
\description{
Build an blue row
}
\examples{
if (interactive()) {
 library(blueR)
 blueRow(
  blueColumn(
    width = 6,
    blueH1(
      display = 3, 
      "BlueR, HTML static template for R", 
      htmltools::span("completed with examples")
    ) \%>\% blueTextColor(color = "white"),
    blueLead(
      "Blue is a great free UI package based on Bootstrap 
              4 that includes the most important components and features"
    ) \%>\% blueTextColor(color = "white")
  ),
  blueColumn(
    width = 6,
    blueImage(
      src = "inst/images/imac.svg",
      floating = TRUE
    ) \%>\% bluePersp(side = "right")
    \%>\% blueBlur()
  )
 )
}


}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
