% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueSection.R
\name{blueSection}
\alias{blueSection}
\title{Create a Boostrap 4 section container}
\usage{
blueSection(
  ...,
  size = NULL,
  status = "default",
  gradient = FALSE,
  separator = FALSE,
  separator_color = "secondary",
  shape = FALSE,
  cascade = FALSE
)
}
\arguments{
\item{...}{Any UI element.}

\item{size}{Section size. NULL by default.}

\item{status}{Section status. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}

\item{gradient}{Section gradient effect. FALSE by default.}

\item{separator}{Section bottom separator. FALSE by default.}

\item{separator_color}{Separator color. "secondary" by default. See \url{https://demos.creative-tim.com/blue-design-system/docs/foundation/colors.html}.}

\item{shape}{Whether to include shape in the background. FALSE by default.}

\item{cascade}{Whether to enable a cascade effect. FALSE by default.}
}
\description{
Build an blue section container
}
\note{
Do not use with the blueDash package. Only for blueR.
}
\examples{
if (interactive()) {
 library(blueR)
 blueSection(
  size = "lg",
  status = "default",
  gradient = TRUE,
  separator = TRUE,
  separator_color = "white",
  shape = TRUE,
  blueColumn(
    blueRow(
      blueColumn(
        width = 6,
        blueH1(
          display = 3, 
          "BlueR, HTML static template for R", 
          htmltools::span("completed with examples")
        ) \%>\% blueTextColor(color = "white"),
        blueLead(
          "Blue is a great free UI package based on Bootstrap 
          4 that includes the most important components and features"
        ) \%>\% blueTextColor(color = "white")
        ),
      blueColumn(
        width = 6,
        blueImage(
          src = "inst/images/imac.svg",
          floating = TRUE
        ) \%>\% bluePersp(side = "right")
        \%>\% blueBlur()
      )
      )
  )
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
