% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueButton.R
\name{blueSocialButton}
\alias{blueSocialButton}
\title{Create a Boostrap 4 blue social button}
\usage{
blueSocialButton(src = NULL, status = "default", icon = NULL)
}
\arguments{
\item{src}{Button external link.}

\item{status}{Button color. See \url{https://demos.creative-tim.com/blue-design-system/docs/components/buttons.html}.}

\item{icon}{Button icon. Expect \link[shiny]{icon}.}
}
\description{
Build an blue social button
}
\examples{
if(interactive()){
 library(blueR)
 blueSocialButton(
  src = "http://rinterface.com",
  status = "danger",
  icon = icon("facebook")
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
