% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cycles_from_peaks}
\alias{cycles_from_peaks}
\title{Make cycle labels}
\usage{
cycles_from_peaks(peaks)
}
\arguments{
\item{peaks}{the list of turning points from peaksign}
}
\value{
list from 1 to nCycles, each repeated by the
  length of the cycle and the whole padded to the length
  of the original series.
}
\description{
Create a cycle labels, 1 to nCycles, given a list of
turning points found using peaksign
The testing direction (i.e. what is a peak)
is found by checking sign of the first peak
The output can be used for grouping a \code{data.table} by cycle
}
