% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{label_cycles}
\alias{label_cycles}
\title{Label cycles & segments}
\usage{
label_cycles(series, span = 3)
}
\arguments{
\item{series}{the list/vector of data to search for peaks.}

\item{span}{size of window to use when looking for peaks. Defaults to 3.
Larger spans are less sensitive to noise, but effectively smooth the series.}
}
\value{
a list of cycle numbers cycle = 1:n.cycles
  and segments seg = rep(c("load","unload"))
  both padded to the length of the original series
}
\description{
Given a data series, break into cycles (at each trough)
break each cycle into load (trough -> peak) and
unload (peak -> trough) segments, using turning points
from peaksign.
As tests start at a trough and go to a peak, the testing
direction is found by checking sign of the first peak/trough.
}
