% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lm_simple}
\alias{lm_simple}
\title{Simplified linear least squares fit}
\usage{
lm_simple(formula, data)
}
\arguments{
\item{formula}{A standard R formula object, passed to /code{lm()}}

\item{data}{Values to fit (must contain the columns in /code{formula})}
}
\value{
A named list (int, slope, p, rsq)
}
\description{
Returns the most useful parts of a basic /code{lm()}
fit to simple univariate (y ~ x) data:
intercept, slope, p-value and r squared
}
