% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{peaks}
\alias{peaks}
\title{Find Peaks}
\usage{
peaks(series, span = 3, do.pad = TRUE)
}
\arguments{
\item{series}{the vector of numbers to find peaks in.}

\item{span}{the window size to use when looking for lower values.}

\item{do.pad}{pad to the same length as the input seris? Default = TRUE.}
}
\value{
a logical vector, TRUE for each peak
}
\description{
Given a series find local peaks by picking values bigger than those on either side
functions posted to the r-help mailing list,
25 November 2005 both by Martin Maechler maechler@stat.math.ethz.ch
(the first function after code by Petr Pikal, in turn after Biran Ripley)
https://stat.ethz.ch/pipermail/r-help/2005-November/083376.html
Assumed to be in the public domain.
}
