% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{peaks2}
\alias{peaks2}
\title{Find Peaks 2}
\usage{
peaks2(series, span = 3)
}
\arguments{
\item{series}{the vector of numbers to find peaks in.}

\item{span}{the window size to use when looking for lower values.}
}
\value{
a logical vector, TRUE for each peak
}
\description{
Given a series find local peaks by picking values bigger than those on either side
functions posted to the r-help mailing list,
23 November 2005 both by Marc Kirchner marc.kirchner@iwr.uni-heidelberg.de
This is the 2004 version by Petr Pikal, after Brian Ripley
=== Petr Pikal in 2004 ===
https://stat.ethz.ch/pipermail/r-help/2005-November/083240.html
Assumed to be in the public domain.
}
