% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{peaksign2}
\alias{peaksign2}
\title{Find Peaks and Troughs 2}
\usage{
peaksign2(series, span = 3, do.pad = TRUE)
}
\arguments{
\item{series}{the vector of numbers to find peaks in.}

\item{span}{the window size to use when looking for lower values.}

\item{do.pad}{should the result be padded to the input series length?
Default is TRUE.}
}
\value{
return (-1 / 0 / 1) if series[i] is ( trough / "normal" / peak )
}
\description{
My version, starting from Brian Ripley's and informed by all the others
}
