% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{segs_from_peaks}
\alias{segs_from_peaks}
\title{Make segment labels}
\usage{
segs_from_peaks(peaks)
}
\arguments{
\item{peaks}{the list of turning points from peaksign}
}
\value{
list of seg = c("load","unload") padded to the length of the original series
}
\description{
Create a list of load/unload segment labels,
given a list of turning points found using peaksign
the first segment is assumed to be "load", followed by "unload" etc.
The output can be used for grouping a \code{data.table} by segment
}
