% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{slope_auto}
\alias{slope_auto}
\title{Bluehill style Automatic Slope}
\usage{
slope_auto(DT, formula)
}
\arguments{
\item{DT}{a \code{data.table} to analyse}

\item{formula}{a standard R formula specifying which columns to use}
}
\value{
A named \code{list(int, slope, p, rsq)} for the segment with maximum slope
}
\description{
Uses the Bluehill Automatic Slope algorithm
1. divide the series into six segments
2. calculate the slope for each segment
3. sum adjacent pairs of slope (1 + 2, 2 + 3 ... 5 + 6)
4. find the pair with the maximum sum
5. return the maximum slope in that pair
This version uses \code{lm_simple} to return other useful statistics
}
