% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{test_direction}
\alias{test_direction}
\title{Find the direction of a test (compressive or tensile)}
\usage{
test_direction(channel)
}
\arguments{
\item{channel}{which channel to test for direction.
Defaults to Load.}
}
\value{
+1 if Tensile, -1 if Compressive
}
\description{
Compressive is defined as -ve Load or Extension (if below crosshead)
so test if the largest value (i.e. the furtherest from the start point)
is greater or less than the start point.
}
