% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trim_slack}
\alias{trim_slack}
\title{Remove intitial slack}
\usage{
trim_slack(DT, formula, lo = 0.02, hi = 0.08)
}
\arguments{
\item{DT}{a \code{data.table} to trim}

\item{formula}{a standard R formula specifying which columns to use as x & y}

\item{lo}{start at  \code{lo * y_max}. DEFAULT is 2%}

\item{hi}{finish at \code{hi * y_max}. DEFAULT is 80%}
}
\value{
a \code{data.table} trimmed to the limits calculated above
}
\description{
Time any initial slack by
1. taking the y data betewwn 2% and 80% of max y (usually Load)
2. fit a straight line
3. calculate the x_intercept (usually Extension)
4. calculate x_max, the x value at max y
5. trim off any data point below x_intercept and x_max
The Bluehill version would use AutoSlope, but this just fits the whole data.
To make the trimmed test start at zero, the first row is subtracted from the
dependent (x) variable as specifed in the input \code{formula} and also
from \code{Time}, if that column exists.
}
