% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_read_study}
\alias{bh_read_study}
\title{Read a set of tests (specimens), given their filenames.
Will add a unique (to this study) testID to each test.}
\usage{
bh_read_study(files)
}
\arguments{
\item{files}{a list of file names/paths to read into the study.}
}
\value{
a list containing two \code{data.table}s
  \code{headers} holding any headers, indexed by testID and
  \code{data}  containing the data rows for all of the files specified.
    Only returns \code{bh_options$min_columns} as all tests need
    to have the same columns to be collected into a table.
  Each specimen has a unique ID in column testID, the same as \code{headers}.
}
\description{
Read a set of tests (specimens), given their filenames.
Will add a unique (to this study) testID to each test.
}
