% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_set_options}
\alias{bh_set_options}
\title{Set __bluer__ options}
\usage{
bh_set_options(raw_regex = ".*RawData.*\\\\.csv$", max_header = 100,
  min_columns = c("Time", "Extension", "Load"),
  id_regex = file.path(".*",
  "(.+).is_.+_RawData.*Specimen_RawData[\\\\_\\\\.](\\\\d+)\\\\.csv"),
  id_parts = c("filename", "sample", "specimen"))
}
\arguments{
\item{raw_regex}{defines what the names of Raw files look like.}

\item{max_header}{max number of rows in a header. Defaults to 100.}

\item{min_columns}{a BlueHill test must contain at least three columns.
Time, Extension & Load}

\item{id_regex}{to extract common identifying parts of a pathname}

\item{id_parts}{the names of the common identifying parts of a pathname}
}
\value{
a named list of options.
    Sets the global \code{bh_options} as a side effect.
}
\description{
Provides an interface to set global options of processing BlueHill files.
}
