% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bluehill_functions.R
\name{bh_slack_correct}
\alias{bh_slack_correct}
\title{Slack Correction}
\usage{
bh_slack_correct(channel)
}
\arguments{
\item{channel}{}
}
\value{
The original series with any initial slack portion trimmed off.
}
\description{
Remove any slack at the start of a test (or cycle)
}
