% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R
\name{bh_slack_correct}
\alias{bh_slack_correct}
\title{Slack Correction}
\usage{
bh_slack_correct(DT)
}
\arguments{
\item{DT}{a \code{data.table} with \code{Load} and \code{Extension} channels to be corrected.}
}
\value{
The original series with any initial slack portion removed.
}
\description{
Remove any slack at the start of a test (or cycle)
}
\details{
The Bluehill version sets the limits at 2\% and 80\% of maximum then uses
AutoSlope to get the estimated slope and sets the new zero to where that line
intersects y = 0.

This version justs fits a single line to the whole range from 2\% to 80\% of maximum
intercept from that to define the zero point. Rows before that point are
removed and the first row subtracted from both \code{Time} and \code{Extension}
so that they start from zero. \code{Load} is unchanged.

Really just a wrapper for \code{trim_slack()}.
}
