% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluehill_functions.R,
%   R/new_bluehill_functions.R
\name{bh_slack_correct}
\alias{bh_slack_correct}
\title{Slack Correction}
\usage{
bh_slack_correct(DT, channel)

bh_slack_correct(DT, channel)
}
\arguments{
\item{DT}{a data.table with \code{Load} and \code{Extension} channels to be corrected.}

\item{channel}{analyse this channel to find starting slack}

\item{DT}{a data.table containing channels to be corrected}
}
\value{
The original series with any initial slack portion removed.

The original series with any initial slack portion trimmed off.
}
\description{
Remove any slack at the start of a test (or cycle)

Remove any slack at the start of a test (or cycle)
}
\details{
The Bluehill version sets the limits at maximum & 2% of maximum then uses
AutoSlope to get the estimated slope and sets the new zero to where that line
intersects y = 0.

This version justs fits a single line to 0.02:0.80 of maximum and uses the
intercept from that to define the zero point. Rows before that point are
removed and the first row subtracted from both \code{Time} and
\code{Extension}, so that they start from zero. \code{Load} is unchanged.

Really just a wrapper for \code{trim_slack()}
}
